add accessibility for inputs & labels, refactor input & form components
