+++ src/components/form/Form.stories.mdx
@@ -1,10 +1,5 @@
 import { Meta, Story, Preview } from "@storybook/addon-docs/blocks";
-import {
-  Form,
-  FormButtons,
-  SubmitButton,
-  CancelButton
-} from ".";
+import { Form, FormButtons, SubmitButton, CancelButton } from ".";
 import { Input, Label, Description } from "../input";
 import { Heading } from "../heading";
 
@@ -15,25 +10,37 @@ import { Heading } from "../heading";
 <Preview>
   <Story name="form">
     <Form>
-      <Heading level={3} marginBottom="spacing04">Personal informations</Heading>
-      <Label name="Name" />
-      <Input name="Name" placeholder="Name" />
+      <Heading level={3} marginBottom="spacing04">
+        Personal informations
+      </Heading>
+      <Label name="Name" htmlFor="name-id" />
+      <Input name="Name" id="name-id" placeholder="Name" />
       <Description>Enter your full name.</Description>
-      <Label name="Email" />
-      <Input name="Email" type="email" placeholder="json@example.com" />
+      <Label name="Email" htmlFor="email-id" />
+      <Input
+        name="Email"
+        id="email-id"
+        type="email"
+        placeholder="json@example.com"
+      />
       <Description>Enter your email.</Description>
-      <Label name="Password" />
-      <Input name="Password" type="password" placeholder="Password" />
+      <Label name="Password" htmlFor="password-id" />
+      <Input
+        name="Password"
+        id="password-id"
+        type="password"
+        placeholder="Password"
+      />
       <Description>Enter your password.</Description>
       <FormButtons>
         <CancelButton
-          onClick={e => {
+          onClick={(e) => {
             e.preventDefault();
             window.history.back();
           }}
         />
         <SubmitButton
-          onClick={e => {
+          onClick={(e) => {
             e.preventDefault();
             window.alert("Clicked!");
           }}
