+++ src/components/input/Input.tsx
@@ -1,178 +1,19 @@
-import React, { FunctionComponent } from "react";
-import styled, { css } from "styled-components";
-import theme from "../../theme";
-import { GnuiContainer, Flex } from "../container";
-import { space, SpaceProps } from "styled-system";
+import * as React from "react";
+import { SpaceProps } from "styled-system";
 import { Spinner } from "../spinner";
 import { SVGIcon } from "../svgicon";
-import { InputGroupProps, InputProps, LabelProps, Status } from "./types";
+import { StyledInputGroupProps, StyledInputProps } from "./types";
+import {
+  Clear,
+  IconsWrap,
+  InputGroup,
+  NotClickable,
+  StyledInput,
+} from "./styles";
 
-const StyledLabel = styled.label<LabelProps>`
-  line-height: ${theme.lineHeight};
-  color: ${theme.color.text.text01};
-  margin-bottom: ${theme.spacing.spacing02};
-  text-transform: ${theme.typography.titleCase};
+type Props = StyledInputGroupProps & StyledInputProps & SpaceProps;
 
-  ${({ required }) =>
-    required &&
-    css`
-      &:after {
-        content: "*";
-        color: ${theme.color.text.error};
-        margin-left: ${theme.spacing.spacing01};
-      }
-    `}
-`;
-
-export const Label: FunctionComponent<LabelProps> = ({
-  name = "Label",
-  required,
-}) => (
-  <Flex>
-    <StyledLabel htmlFor={name} required={required}>
-      {name}
-    </StyledLabel>
-  </Flex>
-);
-
-const setStatusColor = (status: Status) => {
-  if (status === "error") {
-    return css`
-      border: 1px solid ${theme.color.border.error};
-    `;
-  } else if (status === "success") {
-    return css`
-      border: 1px solid ${theme.color.border.input};
-    `;
-  }
-};
-
-export const InputGroup = styled(Flex)<InputGroupProps & SpaceProps>`
-  position: relative;
-  align-items: center;
-  border: 1px solid ${theme.color.border.input};
-  padding: ${theme.spacing.spacing02};
-  border-radius: ${theme.radiusDefault};
-  color: ${theme.color.text.text01};
-  background: ${theme.color.field.default};
-  transition: ${theme.transition};
-  overflow: hidden;
-  &:hover {
-    border: 1px solid ${theme.color.border.focus};
-  }
-  &:focus {
-    outline: 0;
-    border: 1px solid ${theme.color.border.focus};
-  }
-  &:disabled {
-    color: ${theme.color.text.text03};
-    background: ${theme.color.field.disabled};
-    border: 1px solid ${theme.color.border.input};
-    cursor: not-allowed;
-  }
-  border-width: ${(props: InputGroupProps) => props.noBorder && "0"};
-  ${({ status }) =>
-    status &&
-    css`
-      ${setStatusColor(status)}
-    `};
-  ${({ popup }) =>
-    popup &&
-    css`
-      cursor: pointer;
-    `};
-  ${space};
-`;
-
-export const StyledInput = styled.input<InputProps>`
-  border: 0;
-  width: 100%;
-  box-sizing: border-box;
-  background: transparent;
-  font-family: ${theme.typography.fonts.body};
-  font-size: ${(props: InputProps) =>
-    props.small ? theme.fontSizes.sm : theme.fontSizes.md};
-  line-height: ${(props: InputProps) =>
-    props.small ? "1rem" : theme.lineHeight};
-  padding: 0;
-  margin: 0;
-  transition: ${theme.transition};
-  color: ${theme.color.text.text01};
-
-  &::placeholder {
-    color: ${theme.color.text.text03};
-  }
-  &:hover,
-  &:focus {
-    outline: 0;
-    border: 0;
-  }
-  &:disabled {
-    color: ${theme.color.text.text03};
-    background: transparent;
-    border: 0;
-    cursor: not-allowed;
-  }
-
-  ${({ type }) =>
-    type === "search" &&
-    css`
-      margin-left: ${theme.spacing.spacing02};
-    `}
-
-  ${({ popup }) =>
-    popup &&
-    css`
-      cursor: pointer;
-    `}
-`;
-
-const StyledDescription = styled(GnuiContainer)`
-  font-size: ${theme.fontSizes.sm};
-  color: ${theme.color.text.text02};
-  margin: ${theme.spacing.spacing02} 0;
-  width: 100%;
-`;
-type ClearProps = {
-  onClick?: () => void;
-};
-export const IconsWrap = styled.div`
-  position: absolute;
-  top: 0;
-  right: ${theme.spacing.spacing06};
-  bottom: 0;
-  display: flex;
-  align-items: center;
-`;
-
-const Clear = styled.button<ClearProps>`
-  display: flex;
-  align-items: center;
-  padding: ${theme.spacing.spacing02};
-  background: none;
-  border: none;
-  margin-left: auto;
-  outline: none;
-  cursor: pointer;
-`;
-
-const NotClickable = styled.div`
-  pointer-events: none;
-  position: absolute;
-  top: 0;
-  right: ${theme.spacing.spacing02};
-  bottom: 0;
-  display: flex;
-  align-items: center;
-`;
-
-export const Description: FunctionComponent = ({ children }) => (
-  <StyledDescription>{children}</StyledDescription>
-);
-
-export const Input: FunctionComponent<
-  InputGroupProps & InputProps & SpaceProps
-> = React.forwardRef(
+export const Input: React.FC<Props> = React.forwardRef(
   (
     {
       noBorder,
