+++ src/components/input/index.tsx
@@ -1,3 +1,5 @@
 export * from "./Input";
 export * from "./InputUpload";
 export * from "./InputSearch";
+export * from "./Description";
+export * from "./Label";
